refactor: fix class location
